Ejercicio1rehecho

+++ ejerciciosU1/Ejercicio02.py
@@ -1,6 +1,6 @@
 #Ejercicio 2
 horas_trabajo = int(input("Introduce las horas trabajadas: "))
-coste_hora = int(input("Introduce el precio por hora: "))
+coste_hora = float(input("Introduce el precio por hora: "))
 
 importe_total = horas_trabajo * coste_hora
